Minor revision
Previously, I created a second getKRoutes that receives the list of vertices and edges instead of the network file. However, Python does not recognise two functions with the same name. To solve the error, the old function (which receives the network file) was renamed to getKRoutesNetFile and only the new function is called getKRoutes.